+++ src/main/java/lumaceon/mods/clockworkphase2/api/time/ITimeProvider.java
@@ -16,15 +16,15 @@ public interface ITimeProvider extends ITimeConnection
      * @param simulate If true, a simulation will occur, but no time will actually be removed.
      * @return The amount of time that was removed successfully (or would have been removed if simulated).
      */
-    public long extractTime(long maxExtract, boolean simulate);
+    public int extractTime(int maxExtract, boolean simulate);
 
     /**
      * Returns the maximum amount of time that can be stored.
      */
-    public long getMaxCapacity();
+    public int getMaxCapacity();
 
     /**
      * Returns the amount of time that's stored.
      */
-    public long getTimeStored();
+    public int getTimeStored();
 }
